refactor(bigtable): consolidate mock policies
