+++ google/cloud/bigtable/internal/async_row_sampler_test.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/bigtable/internal/async_row_sampler.h"
-#include "google/cloud/bigtable/testing/mock_backoff_policy.h"
+#include "google/cloud/bigtable/testing/mock_policies.h"
 #include "google/cloud/bigtable/testing/mock_response_reader.h"
 #include "google/cloud/bigtable/testing/table_test_fixture.h"
 #include "google/cloud/testing_util/fake_completion_queue_impl.h"
